% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_cv.R
\name{block_cv}
\alias{block_cv}
\title{Use Block Cross-Validation to Evaluate Models}
\usage{
block_cv(data, dayvar = NULL, model, block = 10)
}
\arguments{
\item{data}{A data frame.}

\item{dayvar}{A character string. The name of the variable that represents the day. This is required because this function use dayvar to specify the time point before the test block should not be used to predict the time point after the test block. If dayvar is not specified, then the function will create a new variable called "dayvar", which is a constant, and use it as the dayvar.}

\item{model}{A function. The model to be evaluated. The function should take a data frame as its first argument and return a \code{quadVAR} object. It can be, for example, \code{function(x) quadVAR(x, vars = c("var1", "var2"))}}

\item{block}{An integer. The number of blocks to use in the cross-validation. The default is 10.}
}
\value{
A list of mean squared errors for each model.
}
\description{
This function uses block cross-validation to evaluate a model. The data is split into blocks, and the model is fit on all but one block and evaluated on the remaining block. This process is repeated for each block, and the mean squared error is calculated for each model.
}
