% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-models.R
\name{compare_4_emo}
\alias{compare_4_emo}
\title{Compare estimated model with true model for 4-emotion model}
\usage{
compare_4_emo(model, silent = FALSE)
}
\arguments{
\item{model}{The estimated model, using data simulated from \code{\link[=sim_4_emo]{sim_4_emo()}}, and model estimated using \code{\link[=quadVAR]{quadVAR()}}.}

\item{silent}{Whether to print out the results.}
}
\value{
Silently return data frame with the estimated coefficients and the true coefficients for the main effects and interaction effects, while printing out the results rounded to two digits if \code{silent = FALSE}.
}
\description{
This function compares the estimated model with the true model for the 4-emotion model. It prints out the estimated coefficients and the true coefficients for the main effects and interaction effects.
}
