% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-find-index.R
\name{find_index}
\alias{find_index}
\title{Find index of data that satisfies certain conditions}
\usage{
find_index(data, dayvar, beepvar)
}
\arguments{
\item{data}{A data frame.}

\item{dayvar}{
String indicating assessment day. Adding this argument makes sure that the first measurement of a day is not regressed on the last measurement of the previous day. IMPORTANT: only add this if the data has multiple observations per day.
}

\item{beepvar}{
Optional string indicating assessment beep per day. Adding this argument will cause non-consecutive beeps to be treated as missing!
}
}
\value{
A list of two vectors of indices.
}
\description{
Find index of data that satisfies certain conditions
}
