% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearize.R
\name{get_adj_mat}
\alias{get_adj_mat}
\title{Extract the adjacency matrix from a quadVAR object.}
\usage{
get_adj_mat(model, value)
}
\arguments{
\item{model}{A quadVAR object.}

\item{value}{The \code{actual_value} in the output of \code{\link[=linear_quadVAR_network]{linear_quadVAR_network()}}.}
}
\value{
An adjacency matrix.
}
\description{
Extract the adjacency matrix from a quadVAR object.
}
