% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVAR.R
\name{NVAR}
\alias{NVAR}
\title{Estimate Lag-1 Nested Nonlinear Vector Autoregressive Models}
\usage{
NVAR(data, vars, p, method = c("dredge", "stepAIC"), ...)
}
\arguments{
\item{data}{A \code{tibble}, data.frame, or matrix that represents a time series of vectors, with each row as a time step.}

\item{vars}{A character vector of the variable names used in the model.}

\item{p}{The maximum order of polynomial feature vector.}

\item{method}{One of "dredge" or "stepAIC". "dredge" uses \code{\link[MuMIn:dredge]{MuMIn::dredge()}} and search for all possible submodels, and "stepAIC" uses \code{\link[MASS:stepAIC]{MASS::stepAIC()}} and perform a backward stepwise selection.}

\item{...}{Parameters passed to \code{\link[MuMIn:dredge]{MuMIn::dredge()}} for model selection.}
}
\value{
An \code{NVAR2} object that contains \code{data}, \code{data_td} (a tidy form of \code{tibble} that contains the training data), \code{W_out} (the fitted coefficients), and \code{parameters}.
}
\description{
This function estimate a set of linear and nonlinear autoregressive models, find the best model according to an information criterion, and compare it with the linear AR and VAR models.
}
