% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_plot.R
\name{partial_plot}
\alias{partial_plot}
\title{Make a partial plot of a variable in a model
This function takes a quadVAR model as input, and returns a plot of the partial effect of a variable on the dependent variable (controlling all other variables and the intercept), for higher and lower levels of the moderator variable split by the median.}
\usage{
partial_plot(model, y, x, moderator)
}
\arguments{
\item{model}{A quadVAR model}

\item{y}{The dependent variable}

\item{x}{The variable for which the partial effect is plotted}

\item{moderator}{The moderator variable}
}
\value{
A ggplot object
}
\description{
Make a partial plot of a variable in a model
This function takes a quadVAR model as input, and returns a plot of the partial effect of a variable on the dependent variable (controlling all other variables and the intercept), for higher and lower levels of the moderator variable split by the median.
}
