% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.quadVAR}
\alias{predict.quadVAR}
\title{Predict the values of the dependent variables using the quadVAR model}
\usage{
\method{predict}{quadVAR}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A quadVAR object.}

\item{newdata}{A data frame or tibble containing at least the values of the independent variables, dayvar, and beepvar (if used in model estimation). If NULL, the original data used to fit the model will be used.}

\item{...}{Other arguments passed to the \code{\link[RAMP:predict.RAMP]{RAMP::predict.RAMP()}} function.}
}
\value{
A data frame or tibble containing the predicted values of the dependent variables. If a value cannot be predicted (e.g., because the corresponding previous time point is not in the data), it will be NA.
}
\description{
Predict the values of the dependent variables using the quadVAR model
}
