% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadVAR.R
\name{quadVAR}
\alias{quadVAR}
\alias{print.quadVAR}
\alias{summary.quadVAR}
\alias{coef.quadVAR}
\title{Estimate lag-1 quadratic vector autoregression models}
\usage{
quadVAR(
  data,
  vars,
  penalty = "LASSO",
  tune = "EBIC",
  SIS_options = list(),
  RAMP_options = list()
)

\method{print}{quadVAR}(x, ...)

\method{summary}{quadVAR}(object, ...)

\method{coef}{quadVAR}(object, silent = FALSE, ...)
}
\arguments{
\item{data}{A \code{tibble}, data.frame, or matrix that represents a time series of vectors, with each row as a time step.}

\item{vars}{A character vector of the variable names used in the model.}

\item{penalty}{The penalty used for the linear and regularized VAR models. Possible options include "LASSO", "SCAD", "MCP", with "LASSO" as the default.}

\item{tune}{Tuning parameter selection method. Possible options include "AIC", "BIC", "EBIC", with "EBIC" as the default.}

\item{SIS_options}{A list of other parameters for the \code{\link[SIS:tune.fit]{SIS::tune.fit()}} function. This is used for the regularized  VAR models.}

\item{RAMP_options}{A list of other parameters for the \code{\link[RAMP:RAMP]{RAMP::RAMP()}} function. This is used for the nonlinear quadratic VAR model.}

\item{...}{Not in use.}

\item{object, x}{An \code{quadVAR} object.}

\item{silent}{Whether to print out the results.}
}
\value{
An \code{quadVAR} object that contains the following elements:
\itemize{
\item \code{AR}: A list of linear AR models for each variable.
\item \code{VAR}: A list of regularized VAR models for each variable.
\item \code{fullVAR}: A list of unregularized (full) VAR models for each variable.
\item \code{quadVAR}: A list of regularized nonlinear quadratic VAR models for each variable.
\item \code{fullquadVAR}: A list of unregularized (full) nonlinear quadratic VAR models for each variable.
\item \code{data},\code{vars},\code{penalty},\code{tune},\code{SIS_options},\code{RAMP_options}: The input arguments.
}
}
\description{
This function estimate regularized nonlinear quadratic vector autoregression models with strong hierarchy using the \code{\link[RAMP:RAMP]{RAMP::RAMP()}} algorithm, and also compare it with the linear AR, regularized VAR, and unregularized (full) VAR and quadratic VAR models.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(quadVAR)}: Print the summary for a quadVAR object. See \code{\link[=summary.quadVAR]{summary.quadVAR()}} for details. Note that this function does not automatically print all the coefficients. Use \code{\link[=coef.quadVAR]{coef.quadVAR()}} to extract the coefficients.

\item \code{summary(quadVAR)}: Summary of a quadVAR object. Different IC definitions used by different packages (which differ by a constant) are unified to make them comparable to each other.

\item \code{coef(quadVAR)}: Extract the coefficients from a quadVAR object.

}}
\examples{
set.seed(1614)
data <- sim_4_emo(time = 200, sd = 1)
plot(data[,"x1"])
qV1 <- quadVAR(data, vars = c("x1", "x2", "x3", "x4"), penalty = "LASSO", tune = "EBIC")
summary(qV1)
coef(qV1)


}
