% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearize.R
\name{quadVAR_to_dyn_eqns}
\alias{quadVAR_to_dyn_eqns}
\title{Transform a quadVAR object to a list of dynamic equations.}
\usage{
quadVAR_to_dyn_eqns(model, minus_self = TRUE)
}
\arguments{
\item{model}{A quadVAR object.}

\item{minus_self}{Whether to subtract the term itself from the equation. If \code{TRUE}, the equation will be in the form of (0 =) \code{ ... - X1}; if \code{FALSE}, the equation will be in the form of (X1 = )\code{...}.}
}
\value{
A list of dynamic equations in characters. You can also use \code{\link[rlang:parse_expr]{rlang::parse_expr()}} to parse them into expressions.
}
\description{
Transform a quadVAR object to a list of dynamic equations.
}
