% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_models.R
\name{true_model_4_emo}
\alias{true_model_4_emo}
\title{True model for 4-emotion model}
\usage{
true_model_4_emo(which)
}
\arguments{
\item{which}{Which model to print out. There are four models in total, corresponding to the four variables.}
}
\value{
NULL, but prints out the true model.
}
\description{
This function prints out the true model for the 4-emotion model in the same format as \code{\link[RAMP:RAMP]{RAMP::RAMP()}}, to help users to compare the true model and the estimated model.
}
