% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadVAR.R
\name{quadVAR}
\alias{quadVAR}
\alias{summary.quadVAR}
\title{Estimate Lag-1 Quadratic Vector Autoregression Models}
\usage{
quadVAR(
  data,
  vars,
  penalty = "LASSO",
  tune = "EBIC",
  SIS_options = list(),
  RAMP_options = list()
)

\method{summary}{quadVAR}(object, ...)
}
\arguments{
\item{data}{A \code{tibble}, data.frame, or matrix that represents a time series of vectors, with each row as a time step.}

\item{vars}{A character vector of the variable names used in the model.}

\item{penalty}{The penalty used for the linear and regularized VAR models. Possible options include "LASSO", "SCAD", "MCP", with "LASSO" as the default.}

\item{tune}{Tuning parameter selection method. Possible options include "AIC", "BIC", "EBIC", with "EBIC" as the default.}

\item{SIS_options}{A list of other parameters for the \code{\link[SIS:tune.fit]{SIS::tune.fit()}} function. This is used for the regularized  VAR models.}

\item{RAMP_options}{A list of other parameters for the \code{\link[RAMP:RAMP]{RAMP::RAMP()}} function. This is used for the nonlinear quadratic VAR model.}

\item{object}{An \code{quadVAR} object.}

\item{...}{Not in use.}
}
\value{
An \code{quadVAR} object that contains the following elements:
\itemize{
\item \code{AR}: A list of linear AR models for each variable.
\item \code{VAR}: A list of regularized VAR models for each variable.
\item \code{fullVAR}: A list of unregularized (full) VAR models for each variable.
\item \code{quadVAR}: A list of regularized nonlinear quadratic VAR models for each variable.
\item \code{fullquadVAR}: A list of unregularized (full) nonlinear quadratic VAR models for each variable.
\item \code{data},\code{vars},\code{penalty},\code{tune},\code{SIS_options},\code{RAMP_options}: The input arguments.
}
}
\description{
This function estimate regularized nonlinear quadratic vector autoregression models with strong hierarchy using the \code{\link[RAMP:RAMP]{RAMP::RAMP()}} algorithm, and also compare it with the linear AR, regularized VAR, and unregularized (full) VAR and quadratic VAR models.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary(quadVAR)}: Summary of a quadVAR object. Different IC definitions used by different packages (which differ by a constant) are unified to make them comparable to each other.

}}
